srmech 0.9.0rc7 [phase 2-6/6]: Q-carrier normalization + native Q61 + carrier docs + ship
Completes the v0.9.0rc7 "Q carrier" arc on top of the phase-1 rational→Q flip.

P2 — consumer normalization (~40 sites): flow Q in exact/algebraic contexts;
  collapse to float at genuine FPU boundaries — physical-observable leaf ops
  (qm/sm, qm/bell, qm/octonion, qm/relativistic, qm/gauge, kepler) and iterative
  numeric kernels (laplacian Jacobi/QR/SVD/Fiedler, cascade/compose Kuramoto,
  cascade/matrix_cascades, cascade/hypercomplex_dft, signal_processing taper/window
  helpers). Q has no complex interop by design — the no-interop tripwire that
  catches a leak. Also fixes 6 sigproc window/taper helpers that leaked Q into a
  complex×Q FFT multiply (stft/multitaper/cross_spectral/multirate/dct/sinc_interp/
  heat_kernel) + the laplacian _fsqrt/_fhypot iterative-kernel subroutines, now
  libm-faithful on the non-finite domain (sqrt(+inf)=+inf) the sweeps reach.

P3 — native Q61 C peers: srmech_{sin,cos,atan}_q61 + srmech_{exp,log,sqrt}_q61
  (thin wrappers over the proven cores, exposing the int64 Q61 before the float
  projection) in srmech_trig.c / srmech_explog.c / srmech_sqrt.c + srmech.h; bound
  in _native.py (has_native_trans_q61 gate) and dispatched from rational.py when
  native is present. Byte-exact with the pure cascade. ABI stays 3 (additive).

P4 — test sweep + alias wiring: migrated the float→Q edge/nonfinite/cexp asserts
  across 7 test files; wired the new rational.log into calculus/asymptotic_calculus
  (trig correctly omits it); flipped the 9 ToolEntry return types float→Q in
  tool_schema (cexp/complex_exp stay complex). describe()["tools"]["total"] stays
  317 (no op add/remove). Committed parity test test_native_q61_parity_rc7.py.

P5 — carrier docs: README "Carriers" section + research-notebook §3.39 document the
  five-carrier family (Mat/Vec/HV/Q/Complex128), the ALU-all-the-way / FPU-last-mile
  lens, the per-function A–N cascade reading, and the iterative-FPU-kernel caveat.

P6 — 5-SSOT bumped 0.9.0rc6 → 0.9.0rc7 + CHANGELOG.

numpy-free, MIT, ABI 3, tools.total 317.

+++ docs/srmech/python/srmech/amsc/cascade/compose.py
@@ -542,7 +542,10 @@ def kuramoto_step(
             coupling_sum = 0.0
             for j in range(n):
                 coupling_sum += _rsin(float(theta_list[j]) - theta_i)
-            out.append(theta_i + h * (float(omega_list[i]) + inv_n * coupling_sum))
+            # Kuramoto phase step — float to match the native peer + List[float]
+            # contract (the integrator is an FPU dynamical system; the sin
+            # coupling flowed Q, collapsing at this per-step boundary).
+            out.append(float(theta_i + h * (float(omega_list[i]) + inv_n * coupling_sum)))
         return out
 
     # ── generalised path: validate + dispatch (native or Python) ──────────
@@ -601,7 +604,7 @@ def kuramoto_step(
         f = float(omega_list[i]) + s
         if psi is not None:
             f += ps[i] * _rsin(psi[i] - theta_i)
-        out2.append(theta_i + h * f)
+        out2.append(float(theta_i + h * f))    # float Kuramoto step (see above)
     return out2
